# openapi.json
{
  "components": {
    "schemas": {
...
      "BillingPlanResponse": {
        "description": "Billing plan response.\n\nSchema.org @type: Offer\nMaps: price_monthly_usd \u2192 price, name \u2192 name",
        "properties": {
          "@context": {
            "default": "https://schema.org",
            "title": "@Context",
            "type": "string"
          },
          "@type": {
            "default": "Offer",
            "title": "@Type",
            "type": "string"
          },
          "description": {
            "default": "",
            "description": "Schema.org: description",
            "title": "Description",
            "type": "string"
          },
          "eligible_quantity": {
            "description": "Monthly token quota (Schema.org: eligibleQuantity)",
            "title": "Eligible Quantity",
            "type": "integer"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "title": "Features",
            "type": "array"
          },
          "id": {
            "description": "Plan document ID",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "max_api_keys": {
            "title": "Max Api Keys",
            "type": "integer"
          },
          "name": {
            "description": "Plan name (Schema.org: name)",
            "title": "Name",
            "type": "string"
          },
          "price": {
            "description": "Monthly price in USD (Schema.org: price)",
            "title": "Price",
            "type": "number"
          },
          "price_annual_usd": {
            "title": "Price Annual Usd",
            "type": "number"
          },
          "price_currency": {
            "default": "USD",
            "description": "Schema.org: priceCurrency",
            "title": "Price Currency",
            "type": "string"
          },
          "rate_limit_rpm": {
            "title": "Rate Limit Rpm",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "price",
          "price_annual_usd",
          "eligible_quantity",
          "rate_limit_rpm",
          "max_api_keys",
          "features",
          "is_active"
        ],
        "title": "BillingPlanResponse",
        "type": "object"
      },
      "BillingPortalResponse": {
        "description": "Stripe-style billing portal session.",
        "properties": {
          "expires_at": {
            "description": "Portal session expiry",
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "url": {
            "description": "Hosted billing portal URL",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_at"
        ],
        "title": "BillingPortalResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Standard API error response (RFC 7807-inspired).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable error description",
            "title": "Detail"
          },
          "error": {
            "description": "Machine-readable error code",
            "title": "Error",
            "type": "string"
          },
          "status_code": {
            "description": "HTTP status code",
            "title": "Status Code",
            "type": "integer"
          }
        },
        "required": [
          "error",
          "status_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvoiceLineItemResponse": {
        "description": "Invoice line item.",
        "properties": {
          "amount_usd": {
            "title": "Amount Usd",
            "type": "number"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "unit_price_usd": {
            "title": "Unit Price Usd",
            "type": "number"
          }
        },
        "required": [
          "description",
          "quantity",
          "unit_price_usd",
          "amount_usd"
        ],
        "title": "InvoiceLineItemResponse",
        "type": "object"
      },
      "InvoiceResponse": {
        "description": "Invoice response.\n\nSchema.org @type: Invoice\nMaps: amount_usd \u2192 totalPaymentDue, status \u2192 paymentStatus",
        "properties": {
          "@context": {
            "default": "https://schema.org",
            "title": "@Context",
            "type": "string"
          },
          "@type": {
            "default": "Invoice",
            "title": "@Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "id": {
            "description": "Invoice document ID",
            "title": "Id",
            "type": "string"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineItemResponse"
            },
            "title": "Line Items",
            "type": "array"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "payment_status": {
            "description": "Invoice status (Schema.org: paymentStatus)",
            "title": "Payment Status",
            "type": "string"
          },
          "period_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "period_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "total_payment_due": {
            "description": "Invoice amount in USD (Schema.org: totalPaymentDue)",
            "title": "Total Payment Due",
            "type": "number"
          }
        },
        "required": [
          "id",
          "organization_id",
          "total_payment_due",
          "payment_status",
          "line_items",
          "created_at"
        ],
        "title": "InvoiceResponse",
        "type": "object"
      },
...
      "PaginatedResponse_InvoiceResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "description": "Current page number (1-indexed)",
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "default": 25,
            "description": "Items per page",
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "description": "Total number of items matching the query",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "PaginatedResponse[InvoiceResponse]",
        "type": "object"
      },
...
      "PaymentMethodResponse": {
        "description": "Payment method summary.",
        "properties": {
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "last4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last4"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "type": {
            "description": "card, mobile_money, bank_transfer",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "PaymentMethodResponse",
        "type": "object"
      },
      "SubscriptionResponse": {
        "description": "Subscription response.\n\nSchema.org @type: Order\nMaps: plan \u2192 orderedItem, status \u2192 orderStatus",
        "properties": {
          "@context": {
            "default": "https://schema.org",
            "title": "@Context",
            "type": "string"
          },
          "@type": {
            "default": "Order",
            "title": "@Type",
            "type": "string"
          },
          "cancel_at_period_end": {
            "title": "Cancel At Period End",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_period_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "current_period_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period Start"
          },
          "id": {
            "description": "Subscription document ID",
            "title": "Id",
            "type": "string"
          },
          "order_status": {
            "description": "Subscription status (Schema.org: orderStatus)",
            "title": "Order Status",
            "type": "string"
          },
          "ordered_item": {
            "description": "Plan ID (Schema.org: orderedItem)",
            "title": "Ordered Item",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "payment_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentMethodResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "ordered_item",
          "order_status",
          "organization_id",
          "cancel_at_period_end",
          "created_at"
        ],
        "title": "SubscriptionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "api@shamwari.ai",
      "name": "Nyuchi Web Services",
      "url": "https://shamwari.ai/"
    },
    "description": "Africa's open source AI platform \u2014 API backend.\n\nShamwari AI provides AI inference, model management, and developer APIs purpose-built for African languages and contexts. Built by Nyuchi Africa.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Shamwari AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/billing/invoices": {
      "get": {
        "description": "Returns paginated invoice history with Schema.org Invoice structured data.",
        "operationId": "list_invoices_v1_billing_invoices_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_InvoiceResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List invoices",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/plans": {
      "get": {
        "description": "Returns all active billing plans with Schema.org Offer structured data. Includes pricing, token quotas, rate limits, and feature flags.",
        "operationId": "list_plans_v1_billing_plans_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BillingPlanResponse"
                  },
                  "title": "Response List Plans V1 Billing Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List billing plans",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/portal": {
      "post": {
        "description": "Returns a Stripe-style hosted billing portal URL where the user can manage their card, plan, and invoices. While Stripe is not yet wired up the response points at /pricing on the marketing site so the frontend can already link out from the dashboard.",
        "operationId": "create_billing_portal_session_v1_billing_portal_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPortalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Billing not configured"
          }
        },
        "summary": "Create billing portal session",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/subscription": {
      "get": {
        "description": "Returns the current organization's subscription with Schema.org Order structured data. New accounts default to the Free tier.",
        "operationId": "get_subscription_v1_billing_subscription_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get current subscription",
        "tags": [
          "billing"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.shamwari.ai"
    },
    {
      "description": "Local development",
      "url": "http://localhost:8000"
    }
  ],
...
}